requirements.txt
python-telegram-bot==20.3
flask==2.3.2
flask-restx==1.1.0
requests==2.31.0
requests[socks]==2.31.0
gspread==5.11.0
oauth2client==4.1.3
pytz==2023.3
redis==4.6.0
python-dotenv==1.0.0
pycryptodome==3.18.0
psutil==5.9.5